Potability, pH and Overall Quality Analysis of Drinking Water using Python
In this project we'll deal with a small dataset of water quality. To find hidden insights in the data, we are going to use data analysis techniques with pandas and numpy. The libraries seaborn and matplotlib will be used to visualize the data. A range of exploratory data analysis techniques will be applied in order to provide more clarification regarding the quality of the data.